Prevent ambiguous input arguments for get_variables_metadata (#152)
* Remove ability for get_variables_metadata.py to write more than one output file type (json & csv).
Instead user can invoke it twice to get different formats of output file.
This makes the positional arguments the same as export_dreq_lists_json, which is simpler for users.

* revised --variables_metadata option in export_dreq_lists_json for usage consistency with get_variables_metadata

* parse input lists separated by comma instead of space in get_variables_metadata

* parse input lists separated by comma instead of space in export_dreq_lists_json
also added short versions of optional args, e.g. "-e" for "--experiments"

* removed --version argument from export_dreq_lists_json, because it's no longer needed and also confusing
confusing, because the positional arguments need to be specified even though they're not used
not needed, because the automatic API version warning message says what the installed version is

* updated tests involving get_variables_metadata (removed '-o', comma-separated argument lists)

* updated estimate_dreq_volume -v,-e optional args to use comma-delimited lists

* added UID from Variables table to get_variables_metadata output json file
updated scripts/examples/variables_v1.2.1.json to be consistent with this change